ring/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=ring/requirements.txt pyproject.toml
#
amqp==5.2.0
    # via kombu
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   starlette
    #   watchfiles
billiard==4.2.0
    # via celery
boto3==1.34.14
    # via ring (pyproject.toml)
botocore==1.34.14
    # via
    #   boto3
    #   pynamodb
    #   s3transfer
build==1.0.3
    # via pip-tools
celery==5.3.6
    # via flower
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
dnspython==2.4.2
    # via email-validator
email-validator==2.1.0.post1
    # via fastapi
fastapi[all]==0.104.1
    # via
    #   fastapi
    #   ring (pyproject.toml)
flower==2.0.1
    # via ring (pyproject.toml)
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.25.2
    # via fastapi
humanize==4.9.0
    # via flower
idna==3.6
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.4
    # via celery
markupsafe==2.1.3
    # via jinja2
orjson==3.9.10
    # via fastapi
packaging==23.2
    # via build
pip-tools==7.3.0
    # via ring (pyproject.toml)
prometheus-client==0.19.0
    # via flower
prompt-toolkit==3.0.43
    # via click-repl
pydantic==2.5.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
    #   ring (pyproject.toml)
pydantic-core==2.14.5
    # via pydantic
pydantic-extra-types==2.1.0
    # via fastapi
pydantic-settings==2.1.0
    # via fastapi
pynamodb==5.5.1
    # via ring (pyproject.toml)
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery
python-dotenv==1.0.0
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.6
    # via fastapi
pytz==2023.3.post1
    # via flower
pyyaml==6.0.1
    # via
    #   fastapi
    #   uvicorn
redis==5.0.1
    # via ring (pyproject.toml)
s3transfer==0.10.0
    # via boto3
shortuuid==1.0.11
    # via ring (pyproject.toml)
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.25
    # via ring (pyproject.toml)
starlette==0.27.0
    # via fastapi
tornado==6.4
    # via flower
typing-extensions==4.9.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2023.3
    # via celery
ujson==5.9.0
    # via fastapi
urllib3==2.0.7
    # via botocore
uvicorn[standard]==0.24.0.post1
    # via
    #   fastapi
    #   uvicorn
uvloop==0.19.0
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.21.0
    # via uvicorn
wcwidth==0.2.12
    # via prompt-toolkit
websockets==12.0
    # via uvicorn
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
